ai-task-platform/worker/requirements.txt
pymongo==4.6.1
redis==5.0.1
python-dotenv==1.0.0
dnspython==2.4.2
